+++ tasks/ec2_deploy.js
@@ -10,7 +10,7 @@ var sshCredentials = require('./lib/sshCredentials.js');
 
 module.exports = function(grunt){
 
-    grunt.registerTask('ec2_deploy', function(name){
+    grunt.registerTask('ec2_deploy', 'Deploys to the instance using `rsync`, reloads `pm2` and `nginx`', function(name){
         conf.init(grunt);
 
         if (arguments.length === 0) {