+++ tasks/ec2_run_instance.js
@@ -7,7 +7,7 @@ var conf = require('./lib/conf.js');
 
 module.exports = function(grunt){
 
-    grunt.registerTask('ec2_run_instance', function(name){
+    grunt.registerTask('ec2_run_instance', 'Spins up an EC2 instance, gives a name tag and assigns an IP', function(name){
         conf.init(grunt);
 
         if (arguments.length === 0) {