+++ tasks/ec2_terminate_instance.js
@@ -6,7 +6,7 @@ var conf = require('./lib/conf.js');
 
 module.exports = function(grunt){
 
-    grunt.registerTask('ec2_terminate_instance', function(id){
+    grunt.registerTask('ec2_terminate_instance', 'Terminates an instance', function(id){
         conf.init(grunt);
 
         if (arguments.length === 0) {