+++ tasks/ec2_wait.js
@@ -16,7 +16,7 @@ function wait (fn, duration) {
 
 module.exports = function(grunt){
 
-    grunt.registerTask('ec2_wait', function(name){
+    grunt.registerTask('ec2_wait', 'Waits for an instance to report a public DNS and be accessible through `ssh`', function(name){
         conf.init(grunt);
 
         if (arguments.length === 0) {